Rename node exporter in Flow to `unix`
  - This has been voted internally and `unix` was deemed preferable over
    `node` (which could be confused with nodejs and does not mean
    anything on its own) and `node_exporter`,
    which is too verbose given that we already use the
    `prometheus.exporter` prefix.
  - node_exporter is referenced in the internal documentation, this
    change only affects the component name.
