+++ docs/sources/flow/concepts/configuration_language.md
@@ -69,8 +69,8 @@ formed by merging the component's name (e.g., `local.file`), label (e.g.,
 `password_file`), and export name (e.g., `content`), delimited by period.
 
 For components that don't use labels, like
-`prometheus.exporter.node`, only combine the component name with
-export name: `prometheus.exporter.node.targets`.
+`prometheus.exporter.unix`, only combine the component name with
+export name: `prometheus.exporter.unix.targets`.
 
 ## Blocks
